requirements.txt
paho-mqtt==2.1.0
pyserial==3.5
python-dotenv==1.1.0
